data-import/requirements.txt
Faker>=8.2.0
psycopg2-binary>=2.9.1